Improve swap measuring on systems with dynamic swap.
If a system has a dynamic swap file that can grow infinitelyish, using
too much swap can still be problematic, and that should be visualized
by a color change.

This change adds another measurement of swap usage, and if the new one
is higher than the old one, the new one will be used instead.

What it measures is how large part of the currently used memory has no
room in RAM but must be on disk.  This measurement differs from the
old one in that the new one doesn't care about how much swap is left.