OSX: Fix app bundle copyright information
Now, if you right click Bubblemon.app in the Finder and choose "Get
Info", you'll get the correct copyright information.

The version number in the info box is also kind-of correct, but only
numbers are supported so you won't see if the build was unclean for
example, or if the build is newer than when the tag was set.