OSX: Add TODO about forcing Dock to left or right
Dock performance (which we can't really affect) is about 4x slower
with the Dock at the bottom of the screen compared to when having the
Dock on the side of the screen.

We should ask users when installing if they want their Dock moved to
the left, moved to the right or to not install Bubblemon.