OSX: Upgrade old running Bubblemon if found
If we detect an old (or rather another) Bubblemon running in the Dock,
we replace it with ourselves.

Before this change, we just let it be.

The point is to make it possible for a user to download a new version
and upgrade to that by just launching the new Bubblemon.