Bubbles in the Touch Bar!
The frame resizing needs to be fixed; now when we scale it into
landscape mode on the touch bar we stretch not just the top and bottom
of it, but the right and left edges as well which looks weird.

Also, when we touch Bubblemon, we should launch the Activity Monitor.