TouchBar: Center horizontally
Sometimes we're given too much horizontal space. Don't know what to do
about that, but this change mitigates the effects at least.

I have tried setting the intrinsicContentSize of the BubblemonView, but
that had no visible effect on this problem. Our intrinsic view size was
accessed, but we still got as much width as before.