Try detecting if we're newer than the running Dock
Can be useful when upgrading.

Let's assume the Dock was launched before this build was made.

Given that we restart the Dock as part of installing ourselves we then
know that the Dock is running an older Bubblemon, and we need to
replace it.

Unfortunately this doesn't work, because getting the launch timestamp
for an application seems to only be possible for applications that
were launched by LaunchServices:
https://developer.apple.com/documentation/appkit/nsrunningapplication/1532595-launchdate?language=objc

According to px:
kernel(0)        root
  launchd(1)     root
--> Dock(92561)  johan

Apparently launchd and LaunchServices are two different things:
https://eclecticlight.co/2018/05/22/running-at-startup-when-to-use-a-login-item-or-a-launchagent-launchdaemon/

See also "px launch"; which will show you both launchd and
launchservicesd, demonstrating they are different.