Installer: Reshuffle install order
I got a weird error message while installing the touchbar app.

If we pretend that was because the Dock and its friends were restarting,
then doing the Dock app last should make that error message not happen
again.

The Menu Bar Bubblemon goes first since it's the most normal of the
Bubblemons.